MySQL support
information_schema は ANSI-standard であるので、データ型の変換のみ RDBMS によって切り替えるようにした。

+++ src/main/java/com/example/dataclassgenerator/generator/Column.java
@@ -31,5 +31,5 @@ public class Column {
 	/**
 	 * 長さ.
 	 */
-	private final Integer maxLength;
+	private final Long maxLength;
 }
